apps/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
sqlalchemy==2.0.25
pytest==7.4.4
pytest-asyncio==0.21.1
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pytz==2023.3
